% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otr_expression_plot.R
\name{otr_expression_plot}
\alias{otr_expression_plot}
\title{plot otr expression}
\usage{
otr_expression_plot(model, data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
plot otr expression
}
